+++ src/loop_logic.rs
@@ -119,7 +119,7 @@ impl<'l, Data> LoopHandle<'l, Data> {
     ///
     /// This callback will be called during a dispatching cycle when the event loop has
     /// finished processing all pending events from the sources and becomes idle.
-    pub fn insert_idle<F: FnOnce(&mut Data) + 'l>(&self, callback: F) -> Idle {
+    pub fn insert_idle<'i, F: FnOnce(&mut Data) + 'l + 'i>(&self, callback: F) -> Idle<'i> {
         let mut opt_cb = Some(callback);
         let callback = Rc::new(RefCell::new(Some(move |data: &mut Data| {
             if let Some(cb) = opt_cb.take() {